requirements.txt
anyio==3.6.2
dnspython==2.3.0
fastapi==0.95.0
idna==3.4
motor==3.1.1
pydantic==1.10.7
pymongo==4.3.3
python-decouple==3.8
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
